move sample pages from docs to developer/pages
