+++ developer/pages/.blocklet-page/samples/images/did-connect-v1-demo.png
