+++ developer/pages/pages/samples/auth-demo/index.md
@@ -1,0 +1,9 @@
+---
+title: Auth Demo
+description: Blocklet that use auth by Blocklet Server Auth Service
+layout: my-layout
+---
+
+<SampleInfo sampleName="auth-demo" />
+
+https://github.com/blocklet/auth-demo
