+++ developer/pages/pages/samples/auth-demo/index.zh.md
@@ -1,0 +1,9 @@
+---
+title: Auth Demo
+description: 演示如何使用 Blocklet Server Auth Service 实现认证功能的 Blocklet
+layout: my-layout
+---
+
+<SampleInfo sampleName="auth-demo" />
+
+https://github.com/blocklet/auth-demo
